+++ core/src/main/java/org/jdbi/v3/core/config/ConfigRegistry.java
@@ -45,9 +45,9 @@ public final class ConfigRegistry {
     private ConfigRegistry(ConfigRegistry that) {
         that.configs.forEach((type, config) -> {
             JdbiConfig<?> copy = config.createCopy();
-            copy.setRegistry(this);
             configs.put(type, copy);
         });
+        configs.values().forEach(c -> c.setRegistry(this));
     }
 
     /**