+++ core/src/main/java/org/jdbi/v3/core/mapper/reflect/internal/PojoBuilderUtils.java
@@ -1,3 +1,16 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package org.jdbi.v3.core.mapper.reflect.internal;
 
 import org.jdbi.v3.core.generic.GenericTypes;
@@ -15,14 +28,18 @@ import java.util.List;
 import java.util.Set;
 
 public class PojoBuilderUtils {
+    private static final String[] GETTER_PREFIXES = new String[] {"get", "is"};
+
     static String propertyName(Method m) {
-        final String[] prefixes = new String[] {"get", "is"};
         ColumnName colName = m.getAnnotation(ColumnName.class);
         if (colName != null) {
             return colName.value();
         }
-        final String name = m.getName();
-        for (String prefix : prefixes) {
+        return defaultSetterName(m.getName());
+    }
+
+    static String defaultSetterName(String name) {
+        for (String prefix : GETTER_PREFIXES) {
             if (name.startsWith(prefix)
                 && name.length() > prefix.length()
                 && Character.isUpperCase(name.charAt(prefix.length()))) {
@@ -44,10 +61,10 @@ public class PojoBuilderUtils {
     }
 
     public static MethodHandle findBuilderSetter(final Class<?> builderClass, String name, Method decl, Type type)
-        throws IllegalAccessException, NoSuchMethodException {
+        throws IllegalAccessException {
         final List<NoSuchMethodException> failures = new ArrayList<>();
         final Set<String> names = new LinkedHashSet<>();
-        names.add(decl.getName());
+        names.add(defaultSetterName(decl.getName()));
         names.add(name);
         if (name.length() > 1) {
             final String rest = name.substring(0, 1).toUpperCase() + name.substring(1);