+++ RELEASE_NOTES.md
@@ -2,6 +2,7 @@
   - Kotlin: respect default values in methods when passed null, #1690
   - `Arguments.bindingNullToPrimitivesPermitted` helps you catch
   erroneous binding of `null` to a primitive type
+  - preparable Arguments: re-invoke prepare on each copy, to propagate updated configuration
 
 # 3.12.2
   - Bean binding: ignore getter methods with parameters